linting

+++ custom_components/peaqev/peaqservice/chargecontroller/charger/chargermodel.py
@@ -7,5 +7,5 @@ class ChargerModel:
     disable_current_updates: bool = False
     session_active: bool = False
     latest_charger_call: int = 0
-    lastest_call_off: int = 0
+    lastest_call_off: float = 0
     charger_state_mismatch: bool = False
